extract handleChange function
